chore: fix errors found by testing
