+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-timezone-select",
-  "version": "1.0.9",
+  "version": "1.1.8",
   "description": "Usable, dynamic React Timezone Select",
   "main": "dist/index.js",
   "scripts": {
