+++ src/index.tsx
@@ -2,116 +2,18 @@ import React from "react"
 import Select from "react-select"
 import spacetime from "spacetime"
 import soft from "timezone-soft"
-import type { Props as ReactSelectProps } from "react-select"
+import allTimezones from './timezone-list'
 import "./index.css"
+import type {
+  Props,
+  ITimezone,
+  ITimezoneOption,
+} from './types/timezone'
 
-export type ICustomTimezone = {
-  [key: string]: string
-}
-
-export const i18nTimezones: ICustomTimezone = {
-  "Pacific/Midway": "Midway Island, Samoa",
-  "Pacific/Honolulu": "Hawaii",
-  "America/Juneau": "Alaska",
-  "America/Boise": "Mountain Time",
-  "America/Dawson": "Dawson, Yukon",
-  "America/Chihuahua": "Chihuahua, La Paz, Mazatlan",
-  "America/Phoenix": "Arizona",
-  "America/Chicago": "Central Time",
-  "America/Regina": "Saskatchewan",
-  "America/Mexico_City": "Guadalajara, Mexico City, Monterrey",
-  "America/Belize": "Central America",
-  "America/Detroit": "Eastern Time",
-  "America/Bogota": "Bogota, Lima, Quito",
-  "America/Caracas": "Caracas, La Paz",
-  "America/Santiago": "Santiago",
-  "America/St_Johns": "Newfoundland and Labrador",
-  "America/Sao_Paulo": "Brasilia",
-  "America/Tijuana": "Tijuana",
-  "America/Montevideo": "Montevideo",
-  "America/Argentina/Buenos_Aires": "Buenos Aires, Georgetown",
-  "America/Godthab": "Greenland",
-  "America/Los_Angeles": "Pacific Time",
-  "Atlantic/Azores": "Azores",
-  "Atlantic/Cape_Verde": "Cape Verde Islands",
-  GMT: "UTC",
-  "Europe/London": "Edinburgh, London",
-  "Europe/Dublin": "Dublin",
-  "Europe/Lisbon": "Lisbon",
-  "Africa/Casablanca": "Casablanca, Monrovia",
-  "Atlantic/Canary": "Canary Islands",
-  "Europe/Belgrade": "Belgrade, Bratislava, Budapest, Ljubljana, Prague",
-  "Europe/Sarajevo": "Sarajevo, Skopje, Warsaw, Zagreb",
-  "Europe/Brussels": "Brussels, Copenhagen, Madrid, Paris",
-  "Europe/Amsterdam": "Amsterdam, Berlin, Bern, Rome, Stockholm, Vienna",
-  "Africa/Algiers": "West Central Africa",
-  "Europe/Bucharest": "Bucharest",
-  "Africa/Cairo": "Cairo",
-  "Europe/Helsinki": "Helsinki, Kiev, Riga, Sofia, Tallinn, Vilnius",
-  "Europe/Athens": "Athens, Istanbul, Minsk",
-  "Asia/Jerusalem": "Jerusalem",
-  "Africa/Harare": "Harare, Pretoria",
-  "Europe/Moscow": "Moscow, St. Petersburg, Volgograd",
-  "Asia/Kuwait": "Kuwait, Riyadh",
-  "Africa/Nairobi": "Nairobi",
-  "Asia/Baghdad": "Baghdad",
-  "Asia/Tehran": "Tehran",
-  "Asia/Dubai": "Abu Dhabi, Muscat",
-  "Asia/Baku": "Baku, Tbilisi, Yerevan",
-  "Asia/Kabul": "Kabul",
-  "Asia/Yekaterinburg": "Ekaterinburg",
-  "Asia/Karachi": "Islamabad, Karachi, Tashkent",
-  "Asia/Kolkata": "Chennai, Kolkata, Mumbai, New Delhi",
-  "Asia/Kathmandu": "Kathmandu",
-  "Asia/Dhaka": "Astana, Dhaka",
-  "Asia/Colombo": "Sri Jayawardenepura",
-  "Asia/Almaty": "Almaty, Novosibirsk",
-  "Asia/Rangoon": "Yangon Rangoon",
-  "Asia/Bangkok": "Bangkok, Hanoi, Jakarta",
-  "Asia/Krasnoyarsk": "Krasnoyarsk",
-  "Asia/Shanghai": "Beijing, Chongqing, Hong Kong SAR, Urumqi",
-  "Asia/Kuala_Lumpur": "Kuala Lumpur, Singapore",
-  "Asia/Taipei": "Taipei",
-  "Australia/Perth": "Perth",
-  "Asia/Irkutsk": "Irkutsk, Ulaanbaatar",
-  "Asia/Seoul": "Seoul",
-  "Asia/Tokyo": "Osaka, Sapporo, Tokyo",
-  "Asia/Yakutsk": "Yakutsk",
-  "Australia/Darwin": "Darwin",
-  "Australia/Adelaide": "Adelaide",
-  "Australia/Sydney": "Canberra, Melbourne, Sydney",
-  "Australia/Brisbane": "Brisbane",
-  "Australia/Hobart": "Hobart",
-  "Asia/Vladivostok": "Vladivostok",
-  "Pacific/Guam": "Guam, Port Moresby",
-  "Asia/Magadan": "Magadan, Solomon Islands, New Caledonia",
-  "Asia/Kamchatka": "Kamchatka, Marshall Islands",
-  "Pacific/Fiji": "Fiji Islands",
-  "Pacific/Auckland": "Auckland, Wellington",
-  "Pacific/Tongatapu": "Nuku'alofa",
-}
-
-export type ILabelStyle = "original" | "altName" | "abbrev"
 export enum LabelType {
   ORIGINAL = "original",
   ALTNAME = "altName",
   ABBREV = "abbrev",
-}
-
-export interface ITimezoneOption {
-  value: string
-  label: string
-  abbrev?: string
-  altName?: string
-  offset?: number
-}
-
-export type ITimezone = ITimezoneOption | string
-
-interface Props extends ReactSelectProps {
-  value: ITimezone
-  labelStyle?: ILabelStyle
-  timezones?: ICustomTimezone
 }
 
 const TimezoneSelect = ({
@@ -119,7 +21,7 @@ const TimezoneSelect = ({
   onBlur,
   onChange,
   labelStyle = "original",
-  timezones = i18nTimezones,
+  timezones = allTimezones,
   ...props
 }: Props) => {
   const getOptions = React.useMemo(() => {
@@ -144,10 +46,10 @@ const TimezoneSelect = ({
             label = prefix
             break
           case "altName":
-            label = `${prefix} ${!altName.includes("/") ? `(${altName})` : ""}`
+            label = `${prefix} ${!altName?.includes("/") ? `(${altName})` : ""}`
             break
           case "abbrev":
-            label = `${prefix} ${abbr.length < 5 ? `(${abbr})` : ""}`
+            label = `${prefix} ${abbr?.length < 5 ? `(${abbr})` : ""}`
             break
           default:
             label = `${prefix}`
