+++ src/types/timezone.d.ts
@@ -1,0 +1,23 @@
+import type { Props as ReactSelectProps } from "react-select"
+
+export type ICustomTimezone = {
+  [key: string]: string
+}
+
+export type ILabelStyle = "original" | "altName" | "abbrev"
+
+export interface ITimezoneOption {
+  value: string
+  label: string
+  abbrev?: string
+  altName?: string
+  offset?: number
+}
+
+export type ITimezone = ITimezoneOption | string
+
+interface Props extends ReactSelectProps {
+  value: ITimezone
+  labelStyle?: ILabelStyle
+  timezones?: ICustomTimezone
+}
